refactor: move `simapp.App` to `runtime.AppI`
